# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"PaginationMeta": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"nextPage": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"prevPage": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"lastPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"count",
					"currentPage",
					"nextPage",
					"prevPage",
					"lastPage"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_boolean_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"type": "boolean"
							},
							{
								"items": {
									"type": "boolean"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CategoryType": {
				"enum": [
					"trekking",
					"tours",
					"expeditions",
					"blogs",
					"media"
				],
				"type": "string"
			},
			"CategoryStatus": {
				"enum": [
					"active",
					"draft"
				],
				"type": "string"
			},
			"Category": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/CategoryType"
					},
					"description": {
						"type": "string"
					},
					"itemCount": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/CategoryStatus"
					}
				},
				"required": [
					"id",
					"createdAt",
					"deletedAt",
					"name",
					"slug",
					"type",
					"description",
					"itemCount",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"FaqStatus": {
				"enum": [
					"active",
					"draft"
				],
				"type": "string"
			},
			"Faq": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"question": {
						"type": "string"
					},
					"answer": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/FaqStatus"
					},
					"order": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"createdAt",
					"deletedAt",
					"question",
					"answer",
					"category",
					"status",
					"order"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Faq-Array_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/Faq"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/Faq"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReorderFaqItemDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"order"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReorderFaqsDto": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/ReorderFaqItemDto"
						},
						"type": "array"
					}
				},
				"required": [
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Faq_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Faq"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Faq"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateFaqDto": {
				"properties": {
					"question": {
						"type": "string"
					},
					"answer": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/FaqStatus"
					},
					"order": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"question",
					"answer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateFaqDto": {
				"properties": {
					"question": {
						"type": "string"
					},
					"answer": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/FaqStatus"
					},
					"order": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"BookingPackageType": {
				"enum": [
					"trekking",
					"expedition",
					"tour"
				],
				"type": "string"
			},
			"BookingPaymentStatus": {
				"enum": [
					"paid",
					"deposit_paid",
					"pending",
					"refunded"
				],
				"type": "string"
			},
			"BookingStatus": {
				"enum": [
					"confirmed",
					"in_review",
					"active_trek",
					"completed",
					"cancelled"
				],
				"type": "string"
			},
			"BookingPermitStatus": {
				"enum": [
					"issued",
					"processing",
					"pending_document"
				],
				"type": "string"
			},
			"Booking": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"reference": {
						"type": "string"
					},
					"guestName": {
						"type": "string"
					},
					"guestEmail": {
						"type": "string"
					},
					"guestPhone": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"packageName": {
						"type": "string"
					},
					"packageType": {
						"$ref": "#/components/schemas/BookingPackageType"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"groupSize": {
						"type": "number",
						"format": "double"
					},
					"totalAmountUSD": {
						"type": "number",
						"format": "double"
					},
					"paymentStatus": {
						"$ref": "#/components/schemas/BookingPaymentStatus"
					},
					"bookingStatus": {
						"$ref": "#/components/schemas/BookingStatus"
					},
					"assignedGuide": {
						"type": "string"
					},
					"permitStatus": {
						"$ref": "#/components/schemas/BookingPermitStatus"
					},
					"specialRequests": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"deletedAt",
					"reference",
					"guestName",
					"guestEmail",
					"guestPhone",
					"country",
					"packageName",
					"packageType",
					"startDate",
					"endDate",
					"groupSize",
					"totalAmountUSD",
					"paymentStatus",
					"bookingStatus",
					"assignedGuide",
					"permitStatus",
					"specialRequests"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_Category-Array_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/Category"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/Category"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Category_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Category"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Category"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCategoryDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/CategoryType"
					},
					"description": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/CategoryStatus"
					}
				},
				"required": [
					"name",
					"type",
					"description",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCategoryDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/CategoryType"
					},
					"description": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/CategoryStatus"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Booking-Array_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/Booking"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/Booking"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Booking_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Booking"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Booking"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateBookingDto": {
				"properties": {
					"guestName": {
						"type": "string"
					},
					"guestEmail": {
						"type": "string"
					},
					"guestPhone": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"packageName": {
						"type": "string"
					},
					"packageType": {
						"$ref": "#/components/schemas/BookingPackageType"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"groupSize": {
						"type": "number",
						"format": "double"
					},
					"totalAmountUSD": {
						"type": "number",
						"format": "double"
					},
					"paymentStatus": {
						"$ref": "#/components/schemas/BookingPaymentStatus"
					},
					"bookingStatus": {
						"$ref": "#/components/schemas/BookingStatus"
					},
					"assignedGuide": {
						"type": "string"
					},
					"permitStatus": {
						"$ref": "#/components/schemas/BookingPermitStatus"
					},
					"specialRequests": {
						"type": "string"
					},
					"cfTurnstileToken": {
						"type": "string"
					}
				},
				"required": [
					"guestName",
					"guestEmail",
					"guestPhone",
					"country",
					"packageName",
					"packageType",
					"startDate",
					"endDate",
					"groupSize",
					"totalAmountUSD"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateBookingDto": {
				"properties": {
					"guestName": {
						"type": "string"
					},
					"guestEmail": {
						"type": "string"
					},
					"guestPhone": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"packageName": {
						"type": "string"
					},
					"packageType": {
						"$ref": "#/components/schemas/BookingPackageType"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"groupSize": {
						"type": "number",
						"format": "double"
					},
					"totalAmountUSD": {
						"type": "number",
						"format": "double"
					},
					"paymentStatus": {
						"$ref": "#/components/schemas/BookingPaymentStatus"
					},
					"bookingStatus": {
						"$ref": "#/components/schemas/BookingStatus"
					},
					"assignedGuide": {
						"type": "string"
					},
					"permitStatus": {
						"$ref": "#/components/schemas/BookingPermitStatus"
					},
					"specialRequests": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "rest-nodejs-starter",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/faqs": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Faq-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"FAQs & Consultations"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/FaqStatus"
						}
					},
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Faq_"
								}
							}
						}
					}
				},
				"tags": [
					"FAQs & Consultations"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateFaqDto"
							}
						}
					}
				}
			}
		},
		"/faqs/reorder": {
			"put": {
				"operationId": "Reorder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_boolean_"
								}
							}
						}
					}
				},
				"tags": [
					"FAQs & Consultations"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReorderFaqsDto"
							}
						}
					}
				}
			}
		},
		"/faqs/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Faq_"
								}
							}
						}
					}
				},
				"tags": [
					"FAQs & Consultations"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Faq_"
								}
							}
						}
					}
				},
				"tags": [
					"FAQs & Consultations"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateFaqDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_boolean_"
								}
							}
						}
					}
				},
				"tags": [
					"FAQs & Consultations"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/categories": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Category-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/CategoryType"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Category_"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCategoryDto"
							}
						}
					}
				}
			}
		},
		"/categories/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Category_"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Category_"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCategoryDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_boolean_"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/bookings": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Booking-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BookingStatus"
						}
					},
					{
						"in": "query",
						"name": "packageType",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BookingPackageType"
						}
					},
					{
						"in": "query",
						"name": "paymentStatus",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BookingPaymentStatus"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Booking_"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBookingDto"
							}
						}
					}
				}
			}
		},
		"/bookings/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Booking_"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Booking_"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBookingDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_boolean_"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "https://abc.com/api/v1",
			"description": "prouduction server"
		},
		{
			"url": "http://localhost:4000/api/v1",
			"description": "Development server"
		}
	]
}